Change seed to be 4 bytes/32 bit instead of 2 bytes/16 bit. When the last member of entityState_t fields is not 32 bit, the netcode causes a buffer overwrite that happens to break cl.mSharedMemory pointer. Also, this changes reduces lag *significantly* because entities in the world aren't networking their seed every frame by mistake

+++ codemp/qcommon/q_shared.h
@@ -2247,7 +2247,7 @@ typedef struct entityState_s {
 
 	qboolean		sightsTransition;	// Are we in a sights transition? (Used for player animation)
 	
-	unsigned short	seed;
+	unsigned int	seed;
 } entityState_t;
 
 typedef enum {
